requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
boto3==1.34.115
botocore==1.34.115
certifi==2024.2.2
cffi==2.0.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
clerk-backend-api==4.2.0
click==8.1.7
cryptography==45.0.7
Deprecated==1.3.1
discogs-client==2.3.0
durationpy==0.6
exceptiongroup==1.3.1
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
jikanpy-v4==1.0.2
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==9.2.0
oauthlib==3.2.2
placebo==0.9.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-benedict==0.33.2
python-dateutil==2.9.0.post0
python-fsutil==0.14.1
python-slugify==8.0.4
python3-discogs-client==2.8
PyYAML==6.0.1
rawgpy==1.0.3
requests==2.32.3
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
svix==1.24.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.4
troposphere==4.8.0
types-Deprecated==1.3.1.20260130
types-python-dateutil==2.9.0.20260124
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.6
Werkzeug==3.0.3
wrapt==2.0.1
WTForm==1.0
yarl==1.9.4
zappa==0.59.0
